Fix horizontal scrollbar
The `width` here was causing the search bar to be wider than the
dropdown-content container, which caused a scrollbar to appear.
Confirmed that without specifying width, the search bar is still full
width, and no horizontal scrollbar.

Tested with various screen sizes and on Chrome/Firefox. Also confirmed
that other instances of `d2l-input-search` generally don't specify a
width - some do specify a max-width, but that appears to be for design
purposes (e.g. the search is in a larger container, rather than the
already-constrained dropdown content).